Improve config management.
Remove code that was added to compensate for changes in sd.json when
users upgrade the program. On program restart, default values are loaded
into gv.sd and if an sd.json files is found, the default values are
replaced using those in the file if they exist. this allows for
key-value pairs to be added or removed by simply editing default
dictionary in ospi.py.