Remove authentication from toggling
Although I agree authentication is needed here it breaks the URL authentication method since the URL scheme is non-standard (this is due to the Arduino method and not OSPi specific).

In the urls.py, a regex is used to match the station toggling which makes it difficult to add a third parameter (password) without making things complex. For now, this falls back to the previous method and is inline with the Arduino firmware. I think in the long term a correct syntax should be adopted (ex: /cm?sn1=1&t=0) which allows proper data input parsing.

Thanks