update sip.py
Remove extra space after the shebang and the command.   With this space the shell is unable to figure out the proper interpreter for the file-- user would get shell error "include not found"--when just typing "./sip.py" from the shell.  Does not impact the usual way of starting the script with "python sip.py" so should be harmless.  
Found this when without thinking tried to start the command without specifying an interpreter like the standard startup script does.