Interval display fix
The projected interval programs would not match what actually will run, if the projection was viewed early in the day.  Intervals are based on UTC, and the calculation needs to be run at the time (or simulated time) of the program start, which may be in a different UTC day than when the schedule is viewed. In my case, the schedule displayed properly only after 5PM (7 hours off UTC).